Finished exercise list and details (without image)

+++ app/src/main/java/tmrapps/getinshapeapp/Exercise/Model/ExerciseReposirory.java
@@ -6,9 +6,10 @@ import android.content.Context;
 import android.content.SharedPreferences;
 import android.os.AsyncTask;
 
+import java.util.ArrayList;
 import java.util.List;
 
-import tmrapps.getinshapeapp.ExerciseList.ExerciseListViewModel;
+import tmrapps.getinshapeapp.ExerciseList.Model.ExerciseInCategory;
 import tmrapps.getinshapeapp.ExerciseList.Model.ExerciseListItem;
 import tmrapps.getinshapeapp.ExerciseList.Model.ExerciseListItemRepository;
 import tmrapps.getinshapeapp.GetInShapeApp;
@@ -23,17 +24,54 @@ public class ExerciseReposirory {
 
     MutableLiveData<Exercise> exerciseLiveData;
 
-    MutableLiveData<List<Exercise>> exercisesLiveDT;
+    MutableLiveData<List<ExerciseInCategory>> exercisesByCategoryLiveData;
 
     ExerciseReposirory() {
 
     }
 
+  /*  public void getAll() {
+        synchronized (this) {
+            // Get the last update date
+            long lastUpdateDate = 0;
+            try {
+                lastUpdateDate = GetInShapeApp.getMyContext()
+                        .getSharedPreferences("TAG", Context.MODE_PRIVATE).getLong("lastUpdateDateExercise", 0);
+            } catch (Exception e) {
+
+            }
+
+            ExerciseFirebase.getAllExerciseAndObserve(lastUpdateDate, new ExerciseFirebase.OnExerciseListener() {
+                @Override
+                public void onComplete(List<Exercise> exerciseList) {
+                    updateExerciseDataInLocalStore(exerciseList);
+                }
+
+                @Override
+                public void onComplete() {
+                    // No need to implement
+                }
+            });
+        }
+    }*/
+
     public LiveData<Exercise> getExerciseById(String exerciseId) {
         synchronized (this) {
             if (this.exerciseLiveData == null) {
                 this.exerciseLiveData = new MutableLiveData<Exercise>();
+            }
 
+            GetExerciseByIdTask GetExerciseByIdTask = new GetExerciseByIdTask();
+            GetExerciseByIdTask.execute(exerciseId);
+        }
+
+        return this.exerciseLiveData;
+    }
+
+    public LiveData<List<ExerciseInCategory>> getExerciseByCategory(String categoryId) {
+        if (this.exercisesByCategoryLiveData == null) {
+            this.exercisesByCategoryLiveData = new MutableLiveData<List<ExerciseInCategory>>();
+            synchronized (this) {
                 // Get the last update date
                 long lastUpdateDate = 0;
                 try {
@@ -43,56 +81,72 @@ public class ExerciseReposirory {
 
                 }
 
-                ExerciseFirebase.getExerciseByIdAndObserve(lastUpdateDate, exerciseId, new ExerciseFirebase.OnExerciseListener() {
+                ExerciseFirebase.getAllExerciseAndObserve(lastUpdateDate, categoryId, new ExerciseFirebase.OnExerciseListener() {
                     @Override
-                    public void onComplete(Exercise exercise) {
-                        updateExerciseDataInLocalStore(exercise, exerciseId);
+                    public void onComplete(List<Exercise> exerciseList) {
+                        updateExerciseDataInLocalStore(exerciseList, categoryId);
+                    }
+
+                    @Override
+                    public void onComplete() {
+                        // No need to implement
                     }
                 });
             }
+        } else {
+            GetExerciseListByCategoryTask exerciseListTask = new GetExerciseListByCategoryTask();
+            exerciseListTask.execute(categoryId);
         }
 
-        return this.exerciseLiveData;
+        return this.exercisesByCategoryLiveData;
     }
 
     public void addExercise(Exercise exercise) {
         ExerciseFirebase.addExercise(exercise, new ExerciseFirebase.OnExerciseListener() {
             @Override
-            public void onComplete(Exercise exercise) {
-                //String categoryId, String exerciseName, String exerciseId
+            public void onComplete(List<Exercise> exercise) {
+                // No need to implement
+            }
+
+            @Override
+            public void onComplete() {
                 ExerciseListItemRepository.instance.
                         addExerciseListItemOfCategory(exercise.getCategoryId(), exercise.getName(), exercise.getId());
             }
         });
     }
 
-    private void updateExerciseDataInLocalStore(Exercise data, String exerciseId) {
+    private void updateExerciseDataInLocalStore(List<Exercise> data, String id) {
         GetExerciseListTask exerciseListTask = new GetExerciseListTask();
-        GetExerciseTaskParams getExerciseTatkParams = new GetExerciseTaskParams(exerciseId, data);
+        GetExerciseTaskParams getExerciseTatkParams = new GetExerciseTaskParams(id, data);
         exerciseListTask.execute(getExerciseTatkParams);
     }
 
     private static class GetExerciseTaskParams {
-        String exerciseId;
-        Exercise data;
+        String id;
+        List<Exercise> data;
 
-        GetExerciseTaskParams(String exerciseId, Exercise data) {
-            this.exerciseId = exerciseId;
+        GetExerciseTaskParams(String exerciseId, List<Exercise> data) {
+            this.id = exerciseId;
             this.data = data;
         }
     }
 
-    class GetExerciseListTask extends AsyncTask<GetExerciseTaskParams, String, Exercise> {
+    /**
+     * Get the exercises in category list
+     * Before returning the data the function save it to the db
+     */
+    class GetExerciseListTask extends AsyncTask<GetExerciseTaskParams, String, List<ExerciseInCategory>> {
 
         @Override
-        protected Exercise doInBackground(GetExerciseTaskParams[] lists) {
+        protected List<ExerciseInCategory> doInBackground(GetExerciseTaskParams[] lists) {
 
             if (lists.length > 0) {
 
                 SharedPreferences sharedPreferences = GetInShapeApp.getMyContext()
                         .getSharedPreferences("TAG", Context.MODE_PRIVATE);
-
-                Exercise data = lists[0].data ;
+                List<ExerciseInCategory> exercisesInCategory = new ArrayList<>();
+                List<Exercise> data = lists[0].data ;
                 long lastUpdateDate = 0;
                 try {
                     lastUpdateDate = sharedPreferences.getLong("lastUpdateDateExercise", 0);
@@ -100,21 +154,91 @@ public class ExerciseReposirory {
 
                 }
 
-                if (data != null) {
+                if (data != null && data.size() > 0) {
                     long recentUpdate = lastUpdateDate;
-                    AppLocalStore.db.exerciseDao().insertAll(data);
-                    if (data.getLastUpdateDate() > recentUpdate) {
-                        recentUpdate = data.getLastUpdateDate();
+                    for (Exercise exercise : data) {
+                        AppLocalStore.db.exerciseDao().insertAll(exercise);
+                        if (exercise.getLastUpdateDate() > recentUpdate) {
+                            recentUpdate = exercise.getLastUpdateDate();
+                        }
                     }
+
+                    List<Exercise> exercises = AppLocalStore.db.exerciseDao().getExerciseByCategory(lists[0].id);
+
+                    exercises.forEach(exercise -> {
+                        ExerciseInCategory exerciseInCategory = new ExerciseInCategory();
+                        exerciseInCategory.setId(exercise.getId());
+                        exerciseInCategory.setName(exercise.getName());
+                        exercisesInCategory.add(exerciseInCategory);
+                    });
 
                     SharedPreferences.Editor editor = sharedPreferences.edit();
                     editor.clear();
+                    editor.putLong("lastUpdateDateExercise", recentUpdate);
                     editor.commit();
-                    SharedPreferences.Editor editor1 = sharedPreferences.edit();
-                    editor1.putLong("lastUpdateDateExercise", recentUpdate);
-                    editor1.commit();
                 }
-                Exercise exercise= AppLocalStore.db.exerciseDao().getExerciseDetails(lists[0].exerciseId);
+                return exercisesInCategory;
+            }
+            return null;
+        }
+
+        @Override
+        protected void onPostExecute(List<ExerciseInCategory> exercises) {
+            super.onPostExecute(exercises);
+            exercisesByCategoryLiveData.setValue(exercises);
+        }
+    }
+
+    /**
+     * Get the exercises of category from the db
+     */
+    class GetExerciseListByCategoryTask extends AsyncTask<String, String, List<ExerciseInCategory>> {
+
+        @Override
+        protected List<ExerciseInCategory> doInBackground(String[] lists) {
+
+            if (lists.length > 0) {
+                String categoryId = lists[0] ;
+                List<Exercise> exercises = null;
+                List<ExerciseInCategory> exercisesInCategory = new ArrayList<>();
+                if (categoryId != null) {
+                    exercises= AppLocalStore.db.exerciseDao().getExerciseByCategory(categoryId);
+                }
+
+                exercises.forEach(exercise -> {
+                    ExerciseInCategory exerciseInCategory = new ExerciseInCategory();
+                    exerciseInCategory.setId(exercise.getId());
+                    exerciseInCategory.setName(exercise.getName());
+                    exercisesInCategory.add(exerciseInCategory);
+                });
+
+                return exercisesInCategory;
+            }
+            return null;
+        }
+
+        @Override
+        protected void onPostExecute(List<ExerciseInCategory> exercises) {
+            super.onPostExecute(exercises);
+            exercisesByCategoryLiveData.setValue(exercises);
+        }
+    }
+
+    /**
+     * Get the exercise from the db
+     */
+    class GetExerciseByIdTask extends AsyncTask<String, String, Exercise> {
+
+        @Override
+        protected Exercise doInBackground(String[] lists) {
+
+            if (lists.length > 0) {
+                String exerciseId = lists[0] ;
+                Exercise exercise = null;
+                List<ExerciseInCategory> exercisesInCategory = new ArrayList<>();
+                if (exerciseId != null) {
+                    exercise = AppLocalStore.db.exerciseDao().getExerciseDetails(exerciseId);
+                }
 
                 return exercise;
             }
